Refactor to vite and axios support
