+++ index.html
@@ -1,0 +1,20 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <link rel="icon" href="/favicon.ico" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <meta name="theme-color" content="#000000" />
+    <meta name="description"
+        content="Compilation of notes from routes of MegaTen speedruns, and other games, I've run so far.">
+    <meta name="keywords=" content="Shin Megami Tensei, MegaTen, Speedruns, Sioneus, RTA">
+    <link rel="apple-touch-icon" href="/logo192.png" />
+
+    <title>Sion's Speedrun Notes</title>
+  </head>
+  <body>
+    <noscript>You need to enable JavaScript to run this app.</noscript>
+    <div id="root"></div>
+    <script type="module" src="/src/index.jsx"></script>
+  </body>
+</html>
